Free up pin 1 and 0 after removing RGB LED
Turns out I can't even use the RGB LED because it relies on a Soft Pulse
With Modulation library which happens to be incompatible with the WaveHC
library. This could be for the better since the RGB would have probably
interfered with the actual color it was sensing.

Actually, the RGB LED wouldn't have worked anyway with the Arduino even if
the libraries were compatible. Using the LED would require a gammatable
which is a byte array of size 256. That array with the WaveHC shield
alone brings the global variable usage up to 96% of the limit.
Attempting to add an additional library (SoftPWM) would have been
impossible.